# src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SLSEA Solar Generation Monitoring API",
    "version": "1.0.0",
    "description": "Sri Lanka Sustainable Energy Authority (SLSEA) Real-Time Solar Generation Monitoring & Telemetry Processing REST API.\n\nProvides real-time IoT device telemetry ingestion, administrative hierarchy navigation, operational analytics, district-level aggregation dashboards, conditional caching (ETag/304), and fine-grained Role-Based Access Control (RBAC) with Write-Read security isolation.",
    "contact": {
      "name": "SLSEA Technical Architecture Team",
      "email": "api-support@slsea.gov.lk"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current Environment Base URL"
    },
    {
      "url": "/api/v1",
      "description": "API Version 1 Endpoint"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer authentication token. Supply tokens containing device scopes ('installation:write:{id}') for telemetry ingestion, or user jurisdictional claims ('role': 'national'|'provincial'|'district', 'jurisdiction_id': UUID, 'scopes': ['read:national', 'read:province:{id}', 'read:district:{id}']) for analytical reads."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": ["code", "message", "detail", "timestamp"],
        "properties": {
          "code": {
            "type": "string",
            "example": "RESOURCE_NOT_FOUND",
            "description": "Machine-readable uppercase error identifier code"
          },
          "message": {
            "type": "string",
            "example": "The requested resource could not be found.",
            "description": "Human-readable summary message"
          },
          "detail": {
            "type": "string",
            "nullable": true,
            "example": "SolarInstallation with id '8e27c1b4-6a75-45d3-8be2-7f9999999999' was not found in the database.",
            "description": "Specific technical detail, validation error, or parameter violation"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2026-10-02T04:15:30.000Z",
            "description": "ISO 8601 UTC timestamp of error generation"
          }
        }
      },
...
      "GridSubstation": {
        "type": "object",
        "required": ["id", "name", "capacity_mw", "district_id"],
        "properties": {
          "id": { "type": "string", "format": "uuid", "example": "018f4a12-7b32-7c80-87a1-e0a801234567" },
          "name": { "type": "string", "example": "Kolonnawa Grid Substation" },
          "capacity_mw": { "type": "number", "format": "float", "example": 60.0 },
          "district_id": { "type": "string", "format": "uuid" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "SolarInstallation": {
        "type": "object",
        "required": ["id", "name", "capacity_kw", "installation_type", "meter_id", "grid_substation_id"],
        "properties": {
          "id": { "type": "string", "format": "uuid", "example": "018f4a12-7b32-7c80-87a1-f0a801234567" },
          "name": { "type": "string", "example": "Colombo Port Commercial Solar Array 1" },
          "capacity_kw": { "type": "number", "format": "float", "example": 250.0 },
          "installation_type": { "type": "string", "enum": ["rooftop", "ground_mounted", "floating"], "example": "rooftop" },
          "meter_id": { "type": "string", "example": "MTR-WP-COL-001" },
          "grid_substation_id": { "type": "string", "format": "uuid" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "CompositeInstallation": {
        "type": "object",
        "properties": {
          "installation": {
            "type": "object",
            "properties": {
              "id": { "type": "string", "format": "uuid" },
              "name": { "type": "string" },
              "capacity_kw": { "type": "number" },
              "installation_type": { "type": "string" },
              "meter_id": { "type": "string" },
              "grid_substation": {
                "type": "object",
                "properties": {
                  "id": { "type": "string", "format": "uuid" },
                  "name": { "type": "string" },
                  "capacity_mw": { "type": "number" },
                  "district": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "string", "format": "uuid" },
                      "name": { "type": "string" },
                      "province": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string", "format": "uuid" },
                          "name": { "type": "string" },
                          "code": { "type": "string" }
                        }
                      }
                    }
                  }
                }
              },
              "summary_stats": {
                "type": "object",
                "properties": {
                  "total_readings": { "type": "integer", "example": 672 },
                  "latest_cumulative_energy_kwh": { "type": "number", "example": 8945.2 },
                  "current_power_kw": { "type": "number", "example": 195.4 },
                  "max_power_kw_recorded": { "type": "number", "example": 242.8 },
                  "avg_voltage_v": { "type": "number", "example": 230.15 },
                  "first_reading_timestamp": { "type": "string", "format": "date-time" },
                  "last_reading_timestamp": { "type": "string", "format": "date-time" }
                }
              }
            }
          }
        }
      },
      "LastReading": {
        "type": "object",
        "required": ["installation_id", "meter_id", "last_reading"],
        "properties": {
          "installation_id": { "type": "string", "format": "uuid" },
          "meter_id": { "type": "string", "example": "MTR-WP-COL-001" },
          "last_reading": {
            "type": "object",
            "properties": {
              "id": { "type": "string", "format": "uuid" },
              "timestamp": { "type": "string", "format": "date-time" },
              "power_kw": { "type": "number", "example": 182.5 },
              "energy_kwh": { "type": "number", "example": 8945.2 },
              "voltage_v": { "type": "number", "example": 230.2 }
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/substations/{id}/installations": {
      "get": {
        "tags": ["Hierarchy Navigation"],
        "summary": "Get Solar Installations for a Substation",
        "description": "Scoped collection returning all solar installations interconnected to a specific grid substation.",
        "operationId": "getSubstationInstallations",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Grid Substation UUID",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Installations under substation returned successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "substation": { "$ref": "#/components/schemas/GridSubstation" },
                    "count": { "type": "integer", "example": 12 },
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/SolarInstallation" } }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid UUID format.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "401": {
            "description": "Authentication credentials required.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "403": {
            "description": "Forbidden: User jurisdiction restricts access to this substation.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "404": {
            "description": "Substation not found.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
    "/installations/{id}/composite": {
      "get": {
        "tags": ["Installations"],
        "summary": "Get Composite Installation View",
        "description": "Returns composite installation resource containing metadata, parent substation/district hierarchy, and computed summary telemetry statistics.",
        "operationId": "getCompositeInstallation",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Solar Installation UUID",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Composite installation resource returned.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CompositeInstallation" } } }
          },
          "400": {
            "description": "Invalid UUID format in path.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "401": {
            "description": "Authentication token missing or expired.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "403": {
            "description": "Forbidden: User jurisdiction restricts access to this installation.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "404": {
            "description": "Installation not found.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
    "/installations/{id}/last-reading": {
      "get": {
        "tags": ["Installations"],
        "summary": "Get Single Most Recent Telemetry Reading",
        "description": "Operational read returning the single most recent GenerationReading record for the site as a derived resource.",
        "operationId": "getLastReading",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Solar Installation UUID",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Latest telemetry reading returned.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LastReading" } } }
          },
          "400": {
            "description": "Invalid UUID parameter.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "401": {
            "description": "Authentication token required.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "403": {
            "description": "Forbidden.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "404": {
            "description": "Installation or telemetry readings not found.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
...
  }
}